requirements.txt
langchain == 0.2.1
python-dotenv == 1.0.1
langchain-community == 0.2.1
wikipedia == 1.4.0
pinecone-client == 3.2.2
langchain_pinecone ==  0.1.1   
langchain_groq ==  0.1.4
langchain_cohere == 0.1.5
langchain_core ==  0.2.2
pypdf == 4.2.0
fastapi == 0.111.0
# -e .
